feat: seed over 64k restaurants, update prisma, shift generated files, fix security vulns
