+++ backend/src/dietary-requirement/dietary-requirement.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
-import { Prisma } from 'prisma/generated';
+import { Prisma } from '@prisma/client';
 import { DietaryRequirementDto } from './dto/create-dietary-requirement.dto';
 
 @Injectable()
